refactor: move channels_v2/evaluation_channel.py to channels/evaluation_channel.py
